All logic for comments page has been moved into CommentsContainer.jsx and rendering for other components inside the comments page as well
